// src/services/menu-options/menu-options.types.ts
export interface MenuOption {
  CHILDREN?: MenuOption[]
  CONTENT?: string
  DESCRIPTION?: string
  ICON?: string
  MENU_OPTION_ID: string
  NAME: string
  PARENT_ID?: string
  PATH: string
  TYPE?: 'link' | 'divider' | 'group' | 'item' | 'submenu'
}

export interface OptionWithPermission extends Omit<MenuOption, 'ICON'> {
  PERMISSIONS: {
    PERMISSION_ID: number
    DESCRIPTION: string
    ACTION_ID: number
    ACTION_NAME: string
  }[]
}